+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to the "ev3dev-browser" extension will be documented in this
 
 <!-- Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file. -->
 
+## Unreleased
+
+### Fixed
+- Fixed Windows path separator in "program" in `.vscode/launch.json` not converted to UNIX path.
+
 ## v1.2.0 - 2020-07-20
 ### Changed
 - Initial debug configuration has new example to run current file